+++ sdk/cosmos/examples/create_delete_database.rs
@@ -1,6 +1,6 @@
 use azure_core::HttpClient;
 use azure_cosmos::prelude::*;
-use azure_cosmos::resources::collection::*;
+use collection::*;
 use futures::stream::StreamExt;
 use std::error::Error;
 use std::sync::Arc;
