+++ sdk/cosmos/src/partition_keys.rs
@@ -1,1 +1,4 @@
+/// A collection of keys to partition on
+///
+/// You can learn more about partitioning [here](https://docs.microsoft.com/en-us/azure/cosmos-db/partitioning-overview)
 pub type PartitionKeys = crate::to_json_vector::ToJsonVector;
