+++ sdk/cosmos/src/prelude.rs
@@ -11,18 +11,23 @@
 //! use azure_cosmos::prelude::*;
 //! ```
 
+#[doc(inline)]
 pub use crate::{ConsistencyLevel, CosmosError, PartitionKeys};
 
+#[doc(inline)]
 pub use crate::clients::*;
 
 // Resources
 pub use crate::resources::collection::Offer;
 pub use crate::resources::document::{IndexingDirective, Query};
+#[doc(inline)]
 pub use crate::resources::*;
 
 // Traits
 pub use crate::traits::*;
+#[doc(inline)]
 pub use database::DatabaseName;
+#[doc(inline)]
 pub use user::UserName;
 
 pub use permission::AuthorizationToken;
