+++ sdk/cosmos/src/requests/mod.rs
@@ -1,3 +1,9 @@
+//! Request builder objects for every kind of request.
+//!
+//! These objects are usually created by calling some sort of method on a client. They
+//! then give you the ability to modify your request with certain options and finally
+//! execute the request with the `execute` method.
+
 mod create_collection_builder;
 mod create_database_builder;
 mod create_document_builder;
@@ -42,47 +48,48 @@ mod replace_reference_attachment_builder;
 mod replace_slug_attachment_builder;
 mod replace_stored_procedure_builder;
 mod replace_user_builder;
-pub use self::create_collection_builder::CreateCollectionBuilder;
-pub use self::create_database_builder::CreateDatabaseBuilder;
-pub use self::create_document_builder::CreateDocumentBuilder;
-pub use self::create_or_replace_trigger_builder::CreateOrReplaceTriggerBuilder;
-pub use self::create_or_replace_user_defined_function_builder::CreateOrReplaceUserDefinedFunctionBuilder;
-pub use self::create_permission_builder::CreatePermissionBuilder;
-pub use self::create_reference_attachment_builder::CreateReferenceAttachmentBuilder;
-pub use self::create_slug_attachment_builder::CreateSlugAttachmentBuilder;
-pub use self::create_stored_procedure_builder::CreateStoredProcedureBuilder;
-pub use self::create_user_builder::CreateUserBuilder;
-pub use self::delete_attachment_builder::DeleteAttachmentBuilder;
-pub use self::delete_collection_builder::DeleteCollectionBuilder;
-pub use self::delete_database_builder::DeleteDatabaseBuilder;
-pub use self::delete_document_builder::DeleteDocumentBuilder;
-pub use self::delete_permission_builder::DeletePermissionsBuilder;
-pub use self::delete_stored_procedure_builder::DeleteStoredProcedureBuilder;
-pub use self::delete_trigger_builder::DeleteTriggerBuilder;
-pub use self::delete_user_builder::DeleteUserBuilder;
-pub use self::delete_user_defined_function_builder::DeleteUserDefinedFunctionBuilder;
-pub use self::execute_stored_procedure_builder::ExecuteStoredProcedureBuilder;
-pub use self::get_attachment_builder::GetAttachmentBuilder;
-pub use self::get_collection_builder::GetCollectionBuilder;
-pub use self::get_database_builder::GetDatabaseBuilder;
-pub use self::get_document_builder::GetDocumentBuilder;
-pub use self::get_partition_key_ranges_builder::GetPartitionKeyRangesBuilder;
-pub use self::get_permission_builer::GetPermissionBuilder;
-pub use self::get_user_builder::GetUserBuilder;
-pub use self::list_attachments_builder::ListAttachmentsBuilder;
-pub use self::list_collections_builder::ListCollectionsBuilder;
-pub use self::list_databases_builder::ListDatabasesBuilder;
-pub use self::list_documents_builder::ListDocumentsBuilder;
-pub use self::list_permissions_builder::ListPermissionsBuilder;
-pub use self::list_stored_procedures_builder::ListStoredProceduresBuilder;
-pub use self::list_triggers_builder::ListTriggersBuilder;
-pub use self::list_user_defined_functions_builder::ListUserDefinedFunctionsBuilder;
-pub use self::list_users_builder::ListUsersBuilder;
-pub use self::query_documents_builder::QueryDocumentsBuilder;
-pub use self::replace_collection_builder::ReplaceCollectionBuilder;
-pub use self::replace_document_builder::ReplaceDocumentBuilder;
-pub use self::replace_permission_builder::ReplacePermissionBuilder;
-pub use self::replace_reference_attachment_builder::ReplaceReferenceAttachmentBuilder;
-pub use self::replace_slug_attachment_builder::ReplaceSlugAttachmentBuilder;
-pub use self::replace_stored_procedure_builder::ReplaceStoredProcedureBuilder;
-pub use self::replace_user_builder::ReplaceUserBuilder;
+
+pub use create_collection_builder::CreateCollectionBuilder;
+pub use create_database_builder::CreateDatabaseBuilder;
+pub use create_document_builder::CreateDocumentBuilder;
+pub use create_or_replace_trigger_builder::CreateOrReplaceTriggerBuilder;
+pub use create_or_replace_user_defined_function_builder::CreateOrReplaceUserDefinedFunctionBuilder;
+pub use create_permission_builder::CreatePermissionBuilder;
+pub use create_reference_attachment_builder::CreateReferenceAttachmentBuilder;
+pub use create_slug_attachment_builder::CreateSlugAttachmentBuilder;
+pub use create_stored_procedure_builder::CreateStoredProcedureBuilder;
+pub use create_user_builder::CreateUserBuilder;
+pub use delete_attachment_builder::DeleteAttachmentBuilder;
+pub use delete_collection_builder::DeleteCollectionBuilder;
+pub use delete_database_builder::DeleteDatabaseBuilder;
+pub use delete_document_builder::DeleteDocumentBuilder;
+pub use delete_permission_builder::DeletePermissionsBuilder;
+pub use delete_stored_procedure_builder::DeleteStoredProcedureBuilder;
+pub use delete_trigger_builder::DeleteTriggerBuilder;
+pub use delete_user_builder::DeleteUserBuilder;
+pub use delete_user_defined_function_builder::DeleteUserDefinedFunctionBuilder;
+pub use execute_stored_procedure_builder::ExecuteStoredProcedureBuilder;
+pub use get_attachment_builder::GetAttachmentBuilder;
+pub use get_collection_builder::GetCollectionBuilder;
+pub use get_database_builder::GetDatabaseBuilder;
+pub use get_document_builder::GetDocumentBuilder;
+pub use get_partition_key_ranges_builder::GetPartitionKeyRangesBuilder;
+pub use get_permission_builer::GetPermissionBuilder;
+pub use get_user_builder::GetUserBuilder;
+pub use list_attachments_builder::ListAttachmentsBuilder;
+pub use list_collections_builder::ListCollectionsBuilder;
+pub use list_databases_builder::ListDatabasesBuilder;
+pub use list_documents_builder::ListDocumentsBuilder;
+pub use list_permissions_builder::ListPermissionsBuilder;
+pub use list_stored_procedures_builder::ListStoredProceduresBuilder;
+pub use list_triggers_builder::ListTriggersBuilder;
+pub use list_user_defined_functions_builder::ListUserDefinedFunctionsBuilder;
+pub use list_users_builder::ListUsersBuilder;
+pub use query_documents_builder::QueryDocumentsBuilder;
+pub use replace_collection_builder::ReplaceCollectionBuilder;
+pub use replace_document_builder::ReplaceDocumentBuilder;
+pub use replace_permission_builder::ReplacePermissionBuilder;
+pub use replace_reference_attachment_builder::ReplaceReferenceAttachmentBuilder;
+pub use replace_slug_attachment_builder::ReplaceSlugAttachmentBuilder;
+pub use replace_stored_procedure_builder::ReplaceStoredProcedureBuilder;
+pub use replace_user_builder::ReplaceUserBuilder;
