+++ sdk/cosmos/src/resources/document/mod.rs
@@ -1,3 +1,5 @@
+//! Utilities for interacting with [`Document`]s.
+
 mod document_attributes;
 mod indexing_directive;
 mod query;
