--- sdk/cosmos/src/resources/stored_procedure/mod.rs
@@ -1,25 +1,0 @@
-mod stored_procedure_name;
-
-pub use stored_procedure_name::StoredProcedureName;
-
-pub type Parameters = crate::to_json_vector::ToJsonVector;
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct StoredProcedure {
-    pub id: String,
-    #[serde(rename = "_rid")]
-    pub rid: String,
-    #[serde(rename = "_ts")]
-    pub ts: u64,
-    #[serde(rename = "_self")]
-    pub _self: String,
-    #[serde(rename = "_etag")]
-    pub etag: String,
-    pub body: String,
-}
-
-impl StoredProcedureName for StoredProcedure {
-    fn name(&self) -> &str {
-        &self.id
-    }
-}
