--- sdk/cosmos/src/resources/stored_procedure/stored_procedure_name.rs
@@ -1,15 +1,0 @@
-pub trait StoredProcedureName: std::fmt::Debug {
-    fn name(&self) -> &str;
-}
-
-impl StoredProcedureName for &str {
-    fn name(&self) -> &str {
-        self
-    }
-}
-
-impl StoredProcedureName for String {
-    fn name(&self) -> &str {
-        self.as_ref()
-    }
-}
