+++ sdk/cosmos/src/resources/trigger.rs
@@ -1,13 +1,8 @@
-create_enum!(
-    TriggerOperation,
-    (All, "All"),
-    (Create, "Create"),
-    (Replace, "Replace"),
-    (Delete, "Delete")
-);
+//! Utilities for interacting with [`Trigger`]s.
 
-create_enum!(TriggerType, (Pre, "Pre"), (Post, "Post"));
-
+/// A piece of logic that can be executed before or after creating, deleting, & replacing a document.
+///
+/// You can learn more about triggers [here](https://docs.microsoft.com/en-us/rest/api/cosmos-db/triggers).
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct Trigger {
     #[serde(rename = "_rid")]
@@ -25,3 +20,13 @@ pub struct Trigger {
     pub trigger_type: TriggerType,
     pub body: String,
 }
+
+create_enum!(
+    TriggerOperation,
+    (All, "All"),
+    (Create, "Create"),
+    (Replace, "Replace"),
+    (Delete, "Delete")
+);
+
+create_enum!(TriggerType, (Pre, "Pre"), (Post, "Post"));
