+++ sdk/cosmos/src/resources/user_defined_function.rs
@@ -1,3 +1,6 @@
+/// An extension of the core of the Cosmos DB query language.
+///
+/// You can learn more about user defined functions [here](https://docs.microsoft.com/en-us/rest/api/cosmos-db/user-defined-functions).
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct UserDefinedFunction {
     pub id: String,
