+++ sdk/cosmos/src/responses/mod.rs
@@ -1,3 +1,5 @@
+//! Responses from any call to the Cosmos API.
+
 mod create_collection_response;
 mod create_database_response;
 mod create_document_response;
